Fort details should be doubles instead of longs

+++ src/main/java/com/pokegoapi/api/map/Map.java
@@ -489,7 +489,7 @@ public class Map {
 	 * @param lat the lat
 	 * @return the fort details
 	 */
-	public PokemonFuture<FortDetails> getFortDetailsAsync(String id, long lon, long lat) {
+	public PokemonFuture<FortDetails> getFortDetailsAsync(String id, double lon, double lat) {
 		FortDetailsMessage reqMsg = FortDetailsMessage.newBuilder()
 				.setFortId(id)
 				.setLatitude(lat)
@@ -523,7 +523,7 @@ public class Map {
 	 * @throws LoginFailedException  the login failed exception
 	 * @throws RemoteServerException the remote server exception
 	 */
-	public FortDetails getFortDetails(String id, long lon, long lat)
+	public FortDetails getFortDetails(String id, double lon, double lat)
 			throws LoginFailedException, RemoteServerException {
 		return getFortDetailsAsync(id, lon, lat).toBlocking();
 	}
